Rename and move length counter as a separate file
